Create a RedisClientFacade
